requirements.txt
aiohttp
pyyaml
beautifulsoup4